+++ kurtosis-module/impl/near_module.ts
@@ -4,18 +4,14 @@ import * as log from "loglevel";
 import { addContractHelperDb, ContractHelperDbInfo } from "./services/contract_helper_db";
 import { addContractHelperService, ContractHelperServiceInfo } from "./services/contract_helper";
 import { addIndexer, IndexerInfo } from "./services/indexer";
-import { addExplorerWampService, ExplorerWampInfo } from "./services/explorer_wamp";
 import { addExplorerBackendService, ExplorerBackendInfo } from "./services/explorer_backend";
 import { addExplorerFrontendService, ExplorerFrontendInfo } from "./services/explorer_frontend";
 import { addWallet, WalletInfo } from "./services/wallet";
 import { ExecutableKurtosisModule } from "kurtosis-module-api-lib";
 import { deserializeAndValidateParams } from "./module_io/params_deserializer";
 import { ExecuteResult } from "./module_io/result";
-import { ServiceUrl } from "./service_url";
 
 export type ContainerConfigSupplier = (ipAddr: string) => Result<ContainerConfig, Error>;
-
-const EXPLORER_WAMP_BACKEND_SHARED_SECRET: string = "back";
 
 const EXPLORER_WAMP_BACKEND_FRONTEND_SHARED_NETWORK_NAME: string = "localnet";
 
@@ -68,17 +64,6 @@ export class NearModule implements ExecutableKurtosisModule {
         }
         const contractHelperServiceInfo: ContractHelperServiceInfo = addContractHelperServiceResult.value;
 
-        /*
-        const addExplorerWampResult: Result<ExplorerWampInfo, Error> = await addExplorerWampService(
-            enclaveCtx,
-            EXPLORER_WAMP_BACKEND_SHARED_SECRET
-        );
-        if (addExplorerWampResult.isErr()) {
-            return err(addExplorerWampResult.error);
-        }
-        const explorerWampInfo: ExplorerWampInfo = addExplorerWampResult.value;
-        */
-
         const addExplorerBackendResult: Result<ExplorerBackendInfo, Error> = await addExplorerBackendService(
             enclaveCtx,
             indexerInfo.privateRpcUrl,
@@ -88,11 +73,6 @@ export class NearModule implements ExecutableKurtosisModule {
             contractHelperDbInfo.indexerDb,
             contractHelperDbInfo.analyticsDb,
             contractHelperDbInfo.telemetryDb,
-            /*
-            explorerWampInfo.privateUrl,
-            EXPLORER_WAMP_BACKEND_SHARED_SECRET,
-            EXPLORER_WAMP_BACKEND_FRONTEND_SHARED_NETWORK_NAME,
-            */
         );
         if (addExplorerBackendResult.isErr()) {
             return err(addExplorerBackendResult.error);
@@ -101,14 +81,8 @@ export class NearModule implements ExecutableKurtosisModule {
 
         const addExplorerFrontendResult: Result<ExplorerFrontendInfo, Error> = await addExplorerFrontendService(
             enclaveCtx,
-            // executeParams.backendIpAddress,
             explorerBackendInfo.privateUrl,
             explorerBackendInfo.publicUrl,
-            /*
-            explorerWampInfo.privateUrl,
-            explorerWampInfo.publicUrl,
-            EXPLORER_WAMP_BACKEND_FRONTEND_SHARED_NETWORK_NAME,
-            */
         );
         if (addExplorerFrontendResult.isErr()) {
             return err(addExplorerFrontendResult.error);