+++ docs/changelog.md
@@ -2,6 +2,9 @@
 ### Fixes
 * Upgrade all services to latest versions, as an emergency bugfix for the latest NEAR CLI not working with the images inside Kurtosis
 
+### Removals
+* Removed the WAMP because it's no longer required for the explorer
+
 # 0.5.20
 ### Changes
 * Upgraded Ubuntu machine image in Circle CI configuration to version `ubuntu-2004:202201-02`